+++ src/ContaoManager/Plugin.php
@@ -2,20 +2,31 @@
 
 declare(strict_types=1);
 
-namespace Mailwurm\ImgNaviBundle\ContaoManager;
+/*
+ * This file is part of the img-navi bundle.
+ *
+ * (c) Mathias Ebert
+ *
+ * @license LGPL-3.0-or-later
+ */
+
+namespace Tonsinn\ImgNaviBundle\ContaoManager;
 
 use Contao\CoreBundle\ContaoCoreBundle;
 use Contao\ManagerPlugin\Bundle\BundlePluginInterface;
 use Contao\ManagerPlugin\Bundle\Config\BundleConfig;
 use Contao\ManagerPlugin\Bundle\Parser\ParserInterface;
-use Mailwurm\ImgNaviBundle\ImgNaviBundle;
+use Tonsinn\ImgNaviBundle\TonsinnImgNaviBundle;
 
+/**
+ * Plugin for the Contao Manager.
+ */
 class Plugin implements BundlePluginInterface
 {
     public function getBundles(ParserInterface $parser): array
     {
         return [
-            BundleConfig::create(ImgNaviBundle::class)
+            BundleConfig::create(TonsinnImgNaviBundle::class)
                 ->setLoadAfter([ContaoCoreBundle::class]),
         ];
     }